requirements.txt
python-dotenv>=0.21.1
telethon>=1.26.1
pandas>=1.5.3
openai>=0.26.4
spacy>=3.5.0
srsly>=2.4.5
